Test: MOD 인덱싱 예제1
