--- java/effective-java/effective-java/src/main/java/me/staek/chapter06/item37/Plant.java
@@ -1,63 +1,0 @@
-package me.staek.chapter06.item37;
-
-import java.util.*;
-
-import static java.util.stream.Collectors.groupingBy;
-import static java.util.stream.Collectors.toSet;
-
-class Plant {
-    enum LifeCycle { ANNUAL, PERENNIAL, BIENNIAL }
-
-    final String name;
-    final LifeCycle lifeCycle;
-
-    Plant(String name, LifeCycle lifeCycle) {
-        this.name = name;
-        this.lifeCycle = lifeCycle;
-    }
-
-    @Override public String toString() {
-        return name;
-    }
-
-    public static void main(String[] args) {
-        Plant[] garden = {
-            new Plant("Basil",    LifeCycle.ANNUAL),
-            new Plant("Carroway", LifeCycle.BIENNIAL),
-            new Plant("Dill",     LifeCycle.ANNUAL),
-            new Plant("Lavendar", LifeCycle.PERENNIAL),
-            new Plant("Parsley",  LifeCycle.BIENNIAL),
-            new Plant("Rosemary", LifeCycle.PERENNIAL)
-        };
-
-        @SuppressWarnings("unchecked")
-        Set<Plant>[] plantsByLifeCycleArr =
-                (Set<Plant>[]) new Set[LifeCycle.values().length];
-        for (int i = 0; i < plantsByLifeCycleArr.length; i++)
-            plantsByLifeCycleArr[i] = new HashSet<>();
-        for (Plant p : garden)
-            plantsByLifeCycleArr[p.lifeCycle.ordinal()].add(p);
-        // Print the results
-        for (int i = 0; i < plantsByLifeCycleArr.length; i++) {
-            System.out.printf("%s: %s%n",
-                    LifeCycle.values()[i], plantsByLifeCycleArr[i]);
-        }
-        System.out.println("end first test =====");
-
-
-        Map<LifeCycle, Set<Plant>> plantsByLifeCycle =
-                new EnumMap<>(LifeCycle.class);
-        for (LifeCycle lc : LifeCycle.values())
-            plantsByLifeCycle.put(lc, new HashSet<>());
-        for (Plant p : garden)
-            plantsByLifeCycle.get(p.lifeCycle).add(p);
-        System.out.println(plantsByLifeCycle);
-
-        System.out.println(Arrays.stream(garden)
-                .collect(groupingBy(p -> p.lifeCycle)));
-
-        System.out.println(Arrays.stream(garden)
-                .collect(groupingBy(p -> p.lifeCycle,
-                        () -> new EnumMap<>(LifeCycle.class), toSet())));
-    }
-}
